requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-filter==2.2.0
Pillow==7.1.2